pyproject.toml
[tool.poetry]
name = "researchsummarizerandmore"
version = "0.1.0"
description = ""
authors = ["satyamdas03 <116027621+satyamdas03@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
scholarly = "^1.7.11"
transformers = "^4.44.2"
torch = "^2.4.1"
torchvision = "^0.19.1"
torchaudio = "^2.4.1"
customtkinter = "^5.2.2"
matplotlib = "^3.9.2"
wordcloud = "^1.9.3"
nltk = "^3.9.1"
matplotlib-venn = "^1.1.1"
pybliometrics = "^4.1"
python-dotenv = "^1.0.1"
openai = "0.28"

[tool.poetry.group.dev.dependencies]
beautifulsoup4 = "^4.12.3"
googlesearch-python = "^1.2.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
